requirements.txt
apache-airflow[aws,postgres,celery]
boto3
s3fs
pandas
pyarrow
awswrangler
tqdm
requests
nltk